# _docs/swagger.yaml
basePath: /api/v1/
definitions:
  model.JSONResponseError:
    properties:
      code:
        type: string
      error:
        type: string
    required:
    - code
    type: object
  model.JSONResponseSuccess:
    properties:
      code:
        type: string
      data:
        type: string
    required:
    - code
    type: object
host: localhost:8080
info:
  contact: {}
  title: Good Food Gateway documentation
  version: 1.0.0
paths:
...
  /recipes:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JSONResponseSuccess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.JSONResponseError'
      summary: Create a new recipe
      tags:
      - Recipe
  /recipes/{id}:
    delete:
      parameters:
      - description: ID de la recette
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JSONResponseSuccess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.JSONResponseError'
      summary: Delete a recipe
      tags:
      - Recipe
    get:
      parameters:
      - description: ID de la recette
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JSONResponseSuccess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.JSONResponseError'
      summary: Get recipe by id
      tags:
      - Recipe
    put:
      parameters:
      - description: ID de la recette
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JSONResponseSuccess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.JSONResponseError'
      summary: Update a recipe
      tags:
      - Recipe
  /recipes/shops/{id}:
    get:
      parameters:
      - description: ID du shop
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JSONResponseSuccess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.JSONResponseError'
      summary: Get all recipes for shop
      tags:
      - RecipeAAAAAAA
    post:
      parameters:
      - description: ID du restaurant
        in: path
        name: id
        required: true
        type: string
      - description: ID de la recette
        in: body
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JSONResponseSuccess'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.JSONResponseError'
      summary: Insert a new recipe to a shop
      tags:
      - Recipe
...
swagger: "2.0"